fix: Salesforce CPQ formConfig.json wasn't packaged.
Store formConfig.json as .ts for now.
