+++ src/plugins/salesforce-cpq-config/formConfig.ts
@@ -1,4 +1,21 @@
-{
+export type FormConfig = {
+  [key: string]: {
+    label: string;
+    component: string; // 'tab'
+    id: string;
+    properties: {
+      [key: string]: {
+        label: string;
+        component: string; // 'input' | 'select'
+        type: string; // 'string' | 'boolean' | 'number'
+        name: string;
+        immediatelySave?: boolean;
+      };
+    };
+  };
+};
+
+export const formConfig: FormConfig = {
   "documents": {
     "label": "Documents",
     "component": "tab",
